feat: redis hot cache layer for frequently accessed memories
