+++ tools/agent/structure-rules.yaml
@@ -149,6 +149,8 @@ legacy_hotspots:
       - src/semantic-router/pkg/extproc/req_filter_tools_test.go
     function_checks: relaxed
   - paths:
+      - src/vllm-sr/cli/models.py
+  - paths:
       - src/semantic-router/pkg/extproc/processor_req_body.go
       - src/semantic-router/pkg/extproc/processor_res_body.go
       - src/semantic-router/pkg/promptcompression/compressor_nlp_test.go
